+++ src/main/kotlin/com/github/y0ung3r/gitglobalhookslocator/git/exceptions/ProvidedGitVersionIsInvalidException.kt
@@ -1,0 +1,4 @@
+package com.github.y0ung3r.gitglobalhookslocator.git.exceptions
+
+class ProvidedGitVersionIsInvalidException(version: String)
+    : Exception("The provided version does not conform to git versioning: $version")